Convert to using floats for resources and update manager script.

+++ MapObjects/Resources/Resource.cs
@@ -4,17 +4,17 @@ using System.Collections;
 public class Resource : MapObject {
 	
 	//Inspector fields
-	public int maxResource, initialResource;
+	public float maxResource, initialResource;
 	
 	//Private fields
-	protected int current, max;
+	protected float current, max;
 	
 	//Public properties
-	public int Max {
+	public float Max {
 		get { return max; }
 		private set { max = value; }
 	}
-	public int Current {
+	public float Current {
 		get { return current; }
 		private set { current = value; }
 	}
